Began splitting CommonKernels into multiple files
