+++ platforms/common/src/CommonKernels.cpp
@@ -6,7 +6,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2008-2024 Stanford University and the Authors.      *
+ * Portions copyright (c) 2008-2025 Stanford University and the Authors.      *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -25,6 +25,7 @@
  * -------------------------------------------------------------------------- */
 
 #include "openmm/common/CommonKernels.h"
+#include "openmm/common/CommonKernelUtilities.h"
 #include "openmm/common/ContextSelector.h"
 #include "openmm/common/ExpressionUtilities.h"
 #include "openmm/Context.h"
@@ -35,6 +36,7 @@
 #include "openmm/internal/CustomCompoundBondForceImpl.h"
 #include "openmm/internal/CustomHbondForceImpl.h"
 #include "openmm/internal/CustomManyParticleForceImpl.h"
+#include "openmm/internal/ThreadPool.h"
 #include "openmm/internal/timer.h"
 #include "CommonKernelSources.h"
 #include "lepton/CustomFunction.h"
@@ -54,63 +56,6 @@
 using namespace OpenMM;
 using namespace std;
 using namespace Lepton;
-
-static void setPeriodicBoxArgs(ComputeContext& cc, ComputeKernel kernel, int index) {
-    Vec3 a, b, c;
-    cc.getPeriodicBoxVectors(a, b, c);
-    if (cc.getUseDoublePrecision()) {
-        kernel->setArg(index++, mm_double4(a[0], b[1], c[2], 0.0));
-        kernel->setArg(index++, mm_double4(1.0/a[0], 1.0/b[1], 1.0/c[2], 0.0));
-        kernel->setArg(index++, mm_double4(a[0], a[1], a[2], 0.0));
-        kernel->setArg(index++, mm_double4(b[0], b[1], b[2], 0.0));
-        kernel->setArg(index, mm_double4(c[0], c[1], c[2], 0.0));
-    }
-    else {
-        kernel->setArg(index++, mm_float4((float) a[0], (float) b[1], (float) c[2], 0.0f));
-        kernel->setArg(index++, mm_float4(1.0f/(float) a[0], 1.0f/(float) b[1], 1.0f/(float) c[2], 0.0f));
-        kernel->setArg(index++, mm_float4((float) a[0], (float) a[1], (float) a[2], 0.0f));
-        kernel->setArg(index++, mm_float4((float) b[0], (float) b[1], (float) b[2], 0.0f));
-        kernel->setArg(index, mm_float4((float) c[0], (float) c[1], (float) c[2], 0.0f));
-    }
-}
-
-static bool isZeroExpression(const Lepton::ParsedExpression& expression) {
-    const Lepton::Operation& op = expression.getRootNode().getOperation();
-    if (op.getId() != Lepton::Operation::CONSTANT)
-        return false;
-    return (dynamic_cast<const Lepton::Operation::Constant&>(op).getValue() == 0.0);
-}
-
-static bool usesVariable(const Lepton::ExpressionTreeNode& node, const string& variable) {
-    const Lepton::Operation& op = node.getOperation();
-    if (op.getId() == Lepton::Operation::VARIABLE && op.getName() == variable)
-        return true;
-    for (auto& child : node.getChildren())
-        if (usesVariable(child, variable))
-            return true;
-    return false;
-}
-
-static bool usesVariable(const Lepton::ParsedExpression& expression, const string& variable) {
-    return usesVariable(expression.getRootNode(), variable);
-}
-
-static pair<ExpressionTreeNode, string> makeVariable(const string& name, const string& value) {
-    return make_pair(ExpressionTreeNode(new Operation::Variable(name)), value);
-}
-
-static void flushPeriodically(ComputeContext& cc) {
-#ifdef WIN32
-    // When running on Windows, we periodically flush the queue to keep the UI responsive.
-
-    static double lastTime = getCurrentTime();
-    double currentTime = getCurrentTime();
-    if (currentTime-lastTime > 0.025) {
-        cc.flushQueue();
-        lastTime = currentTime;
-    }
-#endif
-}
 
 void CommonUpdateStateDataKernel::initialize(const System& system) {
 }
@@ -2061,721 +2006,6 @@ void CommonCalcCustomCentroidBondForceKernel::copyParametersToContext(ContextImp
     cc.invalidateMolecules(info, false, true);
 }
 
-class CommonCalcCustomNonbondedForceKernel::ForceInfo : public ComputeForceInfo {
-public:
-    ForceInfo(const CustomNonbondedForce& force) : force(force) {
-        if (force.getNumInteractionGroups() > 0) {
-            groupsForParticle.resize(force.getNumParticles());
-            for (int i = 0; i < force.getNumInteractionGroups(); i++) {
-                set<int> set1, set2;
-                force.getInteractionGroupParameters(i, set1, set2);
-                for (int p : set1)
-                    groupsForParticle[p].insert(2*i);
-                for (int p : set2)
-                    groupsForParticle[p].insert(2*i+1);
-            }
-        }
-    }
-    bool areParticlesIdentical(int particle1, int particle2) {
-        thread_local static vector<double> params1, params2;
-        force.getParticleParameters(particle1, params1);
-        force.getParticleParameters(particle2, params2);
-        for (int i = 0; i < (int) params1.size(); i++)
-            if (params1[i] != params2[i])
-                return false;
-        if (groupsForParticle.size() > 0 && groupsForParticle[particle1] != groupsForParticle[particle2])
-            return false;
-        return true;
-    }
-    int getNumParticleGroups() {
-        return force.getNumExclusions();
-    }
-    void getParticlesInGroup(int index, vector<int>& particles) {
-        int particle1, particle2;
-        force.getExclusionParticles(index, particle1, particle2);
-        particles.resize(2);
-        particles[0] = particle1;
-        particles[1] = particle2;
-    }
-    bool areGroupsIdentical(int group1, int group2) {
-        return true;
-    }
-private:
-    const CustomNonbondedForce& force;
-    vector<set<int> > groupsForParticle;
-};
-
-class CommonCalcCustomNonbondedForceKernel::LongRangePostComputation : public ComputeContext::ForcePostComputation {
-public:
-    LongRangePostComputation(ComputeContext& cc, double& longRangeCoefficient, vector<double>& longRangeCoefficientDerivs, CustomNonbondedForce* force) :
-            cc(cc), longRangeCoefficient(longRangeCoefficient), longRangeCoefficientDerivs(longRangeCoefficientDerivs), force(force) {
-    }
-    double computeForceAndEnergy(bool includeForces, bool includeEnergy, int groups) {
-        if ((groups&(1<<force->getForceGroup())) == 0)
-            return 0;
-        if (!cc.getWorkThread().isCurrentThread())
-            cc.getWorkThread().flush();
-        Vec3 a, b, c;
-        cc.getPeriodicBoxVectors(a, b, c);
-        double volume = a[0]*b[1]*c[2];
-        map<string, double>& derivs = cc.getEnergyParamDerivWorkspace();
-        for (int i = 0; i < longRangeCoefficientDerivs.size(); i++)
-            derivs[force->getEnergyParameterDerivativeName(i)] += longRangeCoefficientDerivs[i]/volume;
-        return longRangeCoefficient/volume;
-    }
-private:
-    ComputeContext& cc;
-    double& longRangeCoefficient;
-    vector<double>& longRangeCoefficientDerivs;
-    CustomNonbondedForce* force;
-};
-
-class CommonCalcCustomNonbondedForceKernel::LongRangeTask : public ComputeContext::WorkTask {
-public:
-    LongRangeTask(ComputeContext& cc, Context& context, CustomNonbondedForceImpl::LongRangeCorrectionData& data,
-                  double& longRangeCoefficient, vector<double>& longRangeCoefficientDerivs, CustomNonbondedForce* force) :
-                        cc(cc), context(context), data(data), longRangeCoefficient(longRangeCoefficient),
-                        longRangeCoefficientDerivs(longRangeCoefficientDerivs), force(force) {
-    }
-    void execute() {
-        CustomNonbondedForceImpl::calcLongRangeCorrection(*force, data, context, longRangeCoefficient, longRangeCoefficientDerivs, cc.getThreadPool());
-    }
-private:
-    ComputeContext& cc;
-    Context& context;
-    CustomNonbondedForceImpl::LongRangeCorrectionData& data;
-    double& longRangeCoefficient;
-    vector<double>& longRangeCoefficientDerivs;
-    CustomNonbondedForce* force;
-};
-
-CommonCalcCustomNonbondedForceKernel::~CommonCalcCustomNonbondedForceKernel() {
-    ContextSelector selector(cc);
-    if (params != NULL)
-        delete params;
-    if (computedValues != NULL)
-        delete computedValues;
-    if (forceCopy != NULL)
-        delete forceCopy;
-}
-
-void CommonCalcCustomNonbondedForceKernel::initialize(const System& system, const CustomNonbondedForce& force) {
-    ContextSelector selector(cc);
-    int forceIndex;
-    for (forceIndex = 0; forceIndex < system.getNumForces() && &system.getForce(forceIndex) != &force; ++forceIndex)
-        ;
-    string prefix = (force.getNumInteractionGroups() == 0 ? "custom"+cc.intToString(forceIndex)+"_" : "");
-
-    // Record parameters and exclusions.
-
-    int numParticles = force.getNumParticles();
-    int paddedNumParticles = cc.getPaddedNumAtoms();
-    int numParams = force.getNumPerParticleParameters();
-    params = new ComputeParameterSet(cc, numParams, paddedNumParticles, "customNonbondedParameters", true);
-    if (force.getNumGlobalParameters() > 0)
-        globals.initialize<float>(cc, force.getNumGlobalParameters(), "customNonbondedGlobals");
-    vector<vector<float> > paramVector(paddedNumParticles, vector<float>(numParams, 0));
-    vector<vector<int> > exclusionList(numParticles);
-    for (int i = 0; i < numParticles; i++) {
-        vector<double> parameters;
-        force.getParticleParameters(i, parameters);
-        paramVector[i].resize(parameters.size());
-        for (int j = 0; j < (int) parameters.size(); j++)
-            paramVector[i][j] = (float) parameters[j];
-        exclusionList[i].push_back(i);
-    }
-    for (int i = 0; i < force.getNumExclusions(); i++) {
-        int particle1, particle2;
-        force.getExclusionParticles(i, particle1, particle2);
-        exclusionList[particle1].push_back(particle2);
-        exclusionList[particle2].push_back(particle1);
-    }
-    params->setParameterValues(paramVector);
-
-    // Record the tabulated functions.
-
-    map<string, Lepton::CustomFunction*> functions;
-    vector<pair<string, string> > functionDefinitions;
-    vector<const TabulatedFunction*> functionList;
-    vector<string> tableTypes;
-    stringstream tableArgs;
-    tabulatedFunctionArrays.resize(force.getNumTabulatedFunctions());
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        functionList.push_back(&force.getTabulatedFunction(i));
-        string name = force.getTabulatedFunctionName(i);
-        tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-        string arrayName = prefix+"table"+cc.intToString(i);
-        functionDefinitions.push_back(make_pair(name, arrayName));
-        functions[name] = cc.getExpressionUtilities().getFunctionPlaceholder(force.getTabulatedFunction(i));
-        int width;
-        vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-        tabulatedFunctionArrays[i].initialize<float>(cc, f.size(), "TabulatedFunction");
-        tabulatedFunctionArrays[i].upload(f);
-        if (force.getNumInteractionGroups() == 0)
-            cc.getNonbondedUtilities().addArgument(ComputeParameterInfo(tabulatedFunctionArrays[i], arrayName, "float", width));
-        if (width == 1)
-            tableTypes.push_back("float");
-        else
-            tableTypes.push_back("float"+cc.intToString(width));
-        tableArgs << ", GLOBAL const float";
-        if (width > 1)
-            tableArgs << width;
-        tableArgs << "* RESTRICT " << arrayName;
-    }
-
-    // Record information for the expressions.
-
-    globalParamNames.resize(force.getNumGlobalParameters());
-    globalParamValues.resize(force.getNumGlobalParameters());
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        globalParamNames[i] = force.getGlobalParameterName(i);
-        globalParamValues[i] = (float) force.getGlobalParameterDefaultValue(i);
-    }
-    if (globals.isInitialized())
-        globals.upload(globalParamValues);
-    bool useCutoff = (force.getNonbondedMethod() != CustomNonbondedForce::NoCutoff);
-    bool usePeriodic = (force.getNonbondedMethod() != CustomNonbondedForce::NoCutoff && force.getNonbondedMethod() != CustomNonbondedForce::CutoffNonPeriodic);
-    Lepton::ParsedExpression energyExpression = Lepton::Parser::parse(force.getEnergyFunction(), functions).optimize();
-    Lepton::ParsedExpression forceExpression = energyExpression.differentiate("r").optimize();
-    map<string, Lepton::ParsedExpression> forceExpressions;
-    forceExpressions["real customEnergy = "] = energyExpression;
-    forceExpressions["tempForce -= "] = forceExpression;
-
-    // Record which per-particle parameters and computed values appear in the energy expression.
-
-    if (force.getNumComputedValues() > 0)
-        computedValues = new ComputeParameterSet(cc, force.getNumComputedValues(), paddedNumParticles, "customNonbondedComputedValues", true);
-    for (int i = 0; i < force.getNumPerParticleParameters(); i++) {
-        string name = force.getPerParticleParameterName(i);
-        if (usesVariable(energyExpression, name+"1") || usesVariable(energyExpression, name+"2")) {
-            paramNames.push_back(name);
-            paramBuffers.push_back(params->getParameterInfos()[i]);
-        }
-    }
-    for (int i = 0; i < force.getNumComputedValues(); i++) {
-        string name, expression;
-        force.getComputedValueParameters(i, name, expression);
-        if (usesVariable(energyExpression, name+"1") || usesVariable(energyExpression, name+"2")) {
-            computedValueNames.push_back(name);
-            computedValueBuffers.push_back(computedValues->getParameterInfos()[i]);
-        }
-    }
-
-    // Create the kernels.
-
-    vector<pair<ExpressionTreeNode, string> > variables;
-    ExpressionTreeNode rnode(new Operation::Variable("r"));
-    variables.push_back(make_pair(rnode, "r"));
-    variables.push_back(make_pair(ExpressionTreeNode(new Operation::Square(), rnode), "r2"));
-    variables.push_back(make_pair(ExpressionTreeNode(new Operation::Reciprocal(), rnode), "invR"));
-    for (int i = 0; i < paramNames.size(); i++) {
-        variables.push_back(makeVariable(paramNames[i]+"1", "((real) "+prefix+"params"+cc.intToString(i+1)+"1)"));
-        variables.push_back(makeVariable(paramNames[i]+"2", "((real) "+prefix+"params"+cc.intToString(i+1)+"2)"));
-    }
-    for (int i = 0; i < computedValueNames.size(); i++) {
-        variables.push_back(makeVariable(computedValueNames[i]+"1", prefix+"values"+cc.intToString(i+1)+"1"));
-        variables.push_back(makeVariable(computedValueNames[i]+"2", prefix+"values"+cc.intToString(i+1)+"2"));
-    }
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        const string& name = force.getGlobalParameterName(i);
-        string value = "globals["+cc.intToString(i)+"]";
-        variables.push_back(makeVariable(name, prefix+value));
-    }
-    for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-        string paramName = force.getEnergyParameterDerivativeName(i);
-        string derivVariable = cc.getNonbondedUtilities().addEnergyParameterDerivative(paramName);
-        Lepton::ParsedExpression derivExpression = energyExpression.differentiate(paramName).optimize();
-        forceExpressions[derivVariable+" += interactionScale*switchValue*"] = derivExpression;
-    }
-    stringstream compute;
-    compute << cc.getExpressionUtilities().createExpressions(forceExpressions, variables, functionList, functionDefinitions, prefix+"temp");
-    map<string, string> replacements;
-    replacements["COMPUTE_FORCE"] = compute.str();
-    replacements["USE_SWITCH"] = (useCutoff && force.getUseSwitchingFunction() ? "1" : "0");
-    if (force.getUseSwitchingFunction()) {
-        // Compute the switching coefficients.
-        
-        replacements["SWITCH_CUTOFF"] = cc.doubleToString(force.getSwitchingDistance());
-        replacements["SWITCH_C3"] = cc.doubleToString(10/pow(force.getSwitchingDistance()-force.getCutoffDistance(), 3.0));
-        replacements["SWITCH_C4"] = cc.doubleToString(15/pow(force.getSwitchingDistance()-force.getCutoffDistance(), 4.0));
-        replacements["SWITCH_C5"] = cc.doubleToString(6/pow(force.getSwitchingDistance()-force.getCutoffDistance(), 5.0));
-    }
-    string source = cc.replaceStrings(CommonKernelSources::customNonbonded, replacements);
-    if (force.getNumInteractionGroups() > 0)
-        initInteractionGroups(force, source, tableTypes);
-    else {
-        cc.getNonbondedUtilities().addInteraction(useCutoff, usePeriodic, true, force.getCutoffDistance(), exclusionList, source, force.getForceGroup(), numParticles > 2000);
-        for (int i = 0; i < paramBuffers.size(); i++)
-            cc.getNonbondedUtilities().addParameter(ComputeParameterInfo(paramBuffers[i].getArray(), prefix+"params"+cc.intToString(i+1),
-                    paramBuffers[i].getComponentType(), paramBuffers[i].getNumComponents()));
-        for (int i = 0; i < computedValueBuffers.size(); i++)
-            cc.getNonbondedUtilities().addParameter(ComputeParameterInfo(computedValueBuffers[i].getArray(), prefix+"values"+cc.intToString(i+1),
-                    computedValueBuffers[i].getComponentType(), computedValueBuffers[i].getNumComponents()));
-        if (globals.isInitialized()) {
-            globals.upload(globalParamValues);
-            cc.getNonbondedUtilities().addArgument(ComputeParameterInfo(globals, prefix+"globals", "float", 1));
-        }
-    }
-    if (force.getNumComputedValues() > 0) {
-        // Create the kernel to calculate computed values.
-
-        stringstream valuesSource, args;
-        for (int i = 0; i < computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string valueName = "values"+cc.intToString(i+1);
-            if (i > 0)
-                args << ", ";
-            args << "GLOBAL " << buffer.getType() << "* RESTRICT global_" << valueName;
-            valuesSource << buffer.getType() << " local_" << valueName << ";\n";
-        }
-        if (force.getNumGlobalParameters() > 0)
-            args << ", GLOBAL const float* globals";
-        for (int i = 0; i < params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            args << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << paramName;
-        }
-        map<string, string> variables;
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++)
-            variables[force.getPerParticleParameterName(i)] = "params"+params->getParameterSuffix(i, "[index]");
-        for (int i = 0; i < force.getNumGlobalParameters(); i++)
-            variables[force.getGlobalParameterName(i)] = "globals["+cc.intToString(i)+"]";
-        for (int i = 0; i < force.getNumComputedValues(); i++) {
-            string name, expression;
-            force.getComputedValueParameters(i, name, expression);
-            variables[name] = "local_values"+computedValues->getParameterSuffix(i);
-            map<string, Lepton::ParsedExpression> valueExpressions;
-            valueExpressions["local_values"+computedValues->getParameterSuffix(i)+" = "] = Lepton::Parser::parse(expression, functions).optimize();
-            valuesSource << cc.getExpressionUtilities().createExpressions(valueExpressions, variables, functionList, functionDefinitions, "value"+cc.intToString(i)+"_temp");
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            string valueName = "values"+cc.intToString(i+1);
-            valuesSource << "global_" << valueName << "[index] = local_" << valueName << ";\n";
-        }
-        map<string, string> replacements;
-        replacements["PARAMETER_ARGUMENTS"] = args.str()+tableArgs.str();
-        replacements["COMPUTE_VALUES"] = valuesSource.str();
-        map<string, string> defines;
-        defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customNonbondedComputedValues, replacements), defines);
-        computedValuesKernel = program->createKernel("computePerParticleValues");
-        for (auto& value : computedValues->getParameterInfos())
-            computedValuesKernel->addArg(value.getArray());
-        if (globals.isInitialized())
-            computedValuesKernel->addArg(globals);
-        for (auto& parameter : params->getParameterInfos())
-            computedValuesKernel->addArg(parameter.getArray());
-        for (auto& function : tabulatedFunctionArrays)
-            computedValuesKernel->addArg(function);
-    }
-    info = new ForceInfo(force);
-    cc.addForce(info);
-    
-    // Record information for the long range correction.
-    
-    if (force.getNonbondedMethod() == CustomNonbondedForce::CutoffPeriodic && force.getUseLongRangeCorrection() && cc.getContextIndex() == 0) {
-        forceCopy = new CustomNonbondedForce(force);
-        longRangeCorrectionData = CustomNonbondedForceImpl::prepareLongRangeCorrection(force, cc.getThreadPool().getNumThreads());
-        cc.addPostComputation(new LongRangePostComputation(cc, longRangeCoefficient, longRangeCoefficientDerivs, forceCopy));
-        hasInitializedLongRangeCorrection = false;
-    }
-    else {
-        longRangeCoefficient = 0.0;
-        hasInitializedLongRangeCorrection = true;
-    }
-}
-
-void CommonCalcCustomNonbondedForceKernel::initInteractionGroups(const CustomNonbondedForce& force, const string& interactionSource, const vector<string>& tableTypes) {
-    // Process groups to form tiles.
-    
-    vector<vector<int> > atomLists;
-    vector<pair<int, int> > tiles;
-    vector<int> tileGroup;
-    vector<vector<int> > duplicateAtomsForGroup;
-    for (int group = 0; group < force.getNumInteractionGroups(); group++) {
-        // Get the list of atoms in this group and sort them.
-        
-        set<int> set1, set2;
-        force.getInteractionGroupParameters(group, set1, set2);
-        vector<int> atoms1, atoms2;
-        atoms1.insert(atoms1.begin(), set1.begin(), set1.end());
-        atoms2.insert(atoms2.begin(), set2.begin(), set2.end());
-        sort(atoms1.begin(), atoms1.end());
-        sort(atoms2.begin(), atoms2.end());
-        duplicateAtomsForGroup.push_back(vector<int>());
-        set_intersection(set1.begin(), set1.end(), set2.begin(), set2.end(),
-                inserter(duplicateAtomsForGroup[group], duplicateAtomsForGroup[group].begin()));
-        sort(duplicateAtomsForGroup[group].begin(), duplicateAtomsForGroup[group].end());
-        
-        // Find how many tiles we will create for this group.
-        
-        int tileWidth = min(min(32, (int) atoms1.size()), (int) atoms2.size());
-        if (tileWidth == 0)
-            continue;
-        int numBlocks1 = (atoms1.size()+tileWidth-1)/tileWidth;
-        int numBlocks2 = (atoms2.size()+tileWidth-1)/tileWidth;
-        
-        // Add the tiles.
-        
-        int firstTile = tiles.size();
-        for (int i = 0; i < numBlocks1; i++)
-            for (int j = 0; j < numBlocks2; j++) {
-                tiles.push_back(make_pair(atomLists.size()+i, atomLists.size()+numBlocks1+j));
-                tileGroup.push_back(group);
-            }
-        
-        // Add the atom lists.
-        
-        for (int i = 0; i < numBlocks1; i++) {
-            vector<int> atoms;
-            int first = i*tileWidth;
-            int last = min((i+1)*tileWidth, (int) atoms1.size());
-            for (int j = first; j < last; j++)
-                atoms.push_back(atoms1[j]);
-            atomLists.push_back(atoms);
-        }
-        for (int i = 0; i < numBlocks2; i++) {
-            vector<int> atoms;
-            int first = i*tileWidth;
-            int last = min((i+1)*tileWidth, (int) atoms2.size());
-            for (int j = first; j < last; j++)
-                atoms.push_back(atoms2[j]);
-            atomLists.push_back(atoms);
-        }
-    }
-    
-    // Build a lookup table for quickly identifying excluded interactions.
-    
-    vector<set<int> > exclusions(force.getNumParticles());
-    for (int i = 0; i < force.getNumExclusions(); i++) {
-        int p1, p2;
-        force.getExclusionParticles(i, p1, p2);
-        exclusions[p1].insert(p2);
-        exclusions[p2].insert(p1);
-    }
-    
-    // Build the exclusion flags for each tile.  While we're at it, filter out tiles
-    // where all interactions are excluded, and sort the tiles by size.
-
-    vector<vector<int> > exclusionFlags(tiles.size());
-    vector<pair<int, int> > tileOrder;
-    for (int tile = 0; tile < tiles.size(); tile++) {
-        bool swapped = false;
-        if (atomLists[tiles[tile].first].size() < atomLists[tiles[tile].second].size()) {
-            // For efficiency, we want the first axis to be the larger one.
-            
-            int swap = tiles[tile].first;
-            tiles[tile].first = tiles[tile].second;
-            tiles[tile].second = swap;
-            swapped = true;
-        }
-        vector<int>& atoms1 = atomLists[tiles[tile].first];
-        vector<int>& atoms2 = atomLists[tiles[tile].second];
-        vector<int>& duplicateAtoms = duplicateAtomsForGroup[tileGroup[tile]];
-        vector<int>& flags = exclusionFlags[tile];
-        flags.resize(atoms1.size(), (int) (1LL<<atoms2.size())-1);
-        int numExcluded = 0;
-        for (int i = 0; i < (int) atoms1.size(); i++) {
-            int a1 = atoms1[i];
-            bool a1IsDuplicate = binary_search(duplicateAtoms.begin(), duplicateAtoms.end(), a1);
-            for (int j = 0; j < (int) atoms2.size(); j++) {
-                int a2 = atoms2[j];
-                bool isExcluded = false;
-                if (a1 == a2 || exclusions[a1].find(a2) != exclusions[a1].end())
-                    isExcluded = true; // This is an excluded interaction.
-                else if ((a1 > a2) == swapped && a1IsDuplicate && binary_search(duplicateAtoms.begin(), duplicateAtoms.end(), a2))
-                    isExcluded = true; // Both atoms are in both sets, so skip duplicate interactions.
-                if (isExcluded) {
-                    flags[i] &= -1-(1<<j);
-                    numExcluded++;
-                }
-            }
-        }
-        if (numExcluded == atoms1.size()*atoms2.size())
-            continue; // All interactions are excluded.
-        tileOrder.push_back(make_pair(-((int)atoms2.size()), tile));
-    }
-    sort(tileOrder.begin(), tileOrder.end());
-    
-    // Merge tiles to get as close as possible to 32 along the first axis of each one.
-    
-    vector<int> tileSetStart;
-    tileSetStart.push_back(0);
-    int tileSetSize = 0;
-    for (int i = 0; i < tileOrder.size(); i++) {
-        int tile = tileOrder[i].second;
-        int size = atomLists[tiles[tile].first].size();
-        if (tileSetSize+size > 32) {
-            tileSetStart.push_back(i);
-            tileSetSize = 0;
-        }
-        tileSetSize += size;
-    }
-    tileSetStart.push_back(tileOrder.size());
-    
-    // Build the data structures.
-    
-    int numTileSets = tileSetStart.size()-1;
-    vector<mm_int4> groupData;
-    for (int tileSet = 0; tileSet < numTileSets; tileSet++) {
-        int indexInTileSet = 0;
-        int minSize = 0;
-        if (cc.getSIMDWidth() < 32) {
-            // We need to include a barrier inside the inner loop, so ensure that all
-            // threads will loop the same number of times.
-            
-            for (int i = tileSetStart[tileSet]; i < tileSetStart[tileSet+1]; i++)
-                minSize = max(minSize, (int) atomLists[tiles[tileOrder[i].second].first].size());
-        }
-        for (int i = tileSetStart[tileSet]; i < tileSetStart[tileSet+1]; i++) {
-            int tile = tileOrder[i].second;
-            vector<int>& atoms1 = atomLists[tiles[tile].first];
-            vector<int>& atoms2 = atomLists[tiles[tile].second];
-            int range = indexInTileSet + ((indexInTileSet+max(minSize, (int) atoms1.size()))<<16);
-            int allFlags = (1<<atoms2.size())-1;
-            for (int j = 0; j < (int) atoms1.size(); j++) {
-                int a1 = atoms1[j];
-                int a2 = (j < atoms2.size() ? atoms2[j] : 0);
-                int flags = (exclusionFlags[tile].size() > 0 ? exclusionFlags[tile][j] : allFlags);
-                groupData.push_back(mm_int4(a1, a2, range, flags<<indexInTileSet));
-            }
-            indexInTileSet += atoms1.size();
-        }
-        for (; indexInTileSet < 32; indexInTileSet++)
-            groupData.push_back(mm_int4(0, 0, minSize<<16, 0));
-    }
-    interactionGroupData.initialize<mm_int4>(cc, groupData.size(), "interactionGroupData");
-    interactionGroupData.upload(groupData);
-    numGroupTiles.initialize<int>(cc, 1, "numGroupTiles");
-
-    // Allocate space for a neighbor list, if necessary.
-
-    if (force.getNonbondedMethod() != CustomNonbondedForce::NoCutoff && groupData.size() > cc.getNumThreadBlocks()) {
-        filteredGroupData.initialize<mm_int4>(cc, groupData.size(), "filteredGroupData");
-        interactionGroupData.copyTo(filteredGroupData);
-        int numTiles = groupData.size()/32;
-        numGroupTiles.upload(&numTiles);
-    }
-    
-    // Create the kernel.
-    
-    hasParamDerivs = (force.getNumEnergyParameterDerivatives() > 0);
-    map<string, string> replacements;
-    replacements["COMPUTE_INTERACTION"] = interactionSource;
-    const string suffixes[] = {"x", "y", "z", "w"};
-    stringstream localData;
-    int localDataSize = 0;
-    for (int i = 0; i < paramBuffers.size(); i++) {
-        localData<<paramBuffers[i].getComponentType()<<" params"<<(i+1)<<";\n";
-        localDataSize += paramBuffers[i].getSize();
-    }
-    for (int i = 0; i < computedValueBuffers.size(); i++) {
-        localData<<computedValueBuffers[i].getComponentType()<<" values"<<(i+1)<<";\n";
-        localDataSize += computedValueBuffers[i].getSize();
-    }
-    replacements["ATOM_PARAMETER_DATA"] = localData.str();
-    stringstream args;
-    for (int i = 0; i < paramBuffers.size(); i++)
-        args<<", GLOBAL const "<<paramBuffers[i].getType()<<"* RESTRICT global_params"<<(i+1);
-    for (int i = 0; i < computedValueBuffers.size(); i++)
-        args<<", GLOBAL const "<<computedValueBuffers[i].getType()<<"* RESTRICT global_values"<<(i+1);
-    for (int i = 0; i < tabulatedFunctionArrays.size(); i++)
-        args << ", GLOBAL const " << tableTypes[i]<< "* RESTRICT table" << i;
-    if (globals.isInitialized())
-        args<<", GLOBAL const float* RESTRICT globals";
-    if (hasParamDerivs)
-        args << ", GLOBAL mixed* RESTRICT energyParamDerivs";
-    replacements["PARAMETER_ARGUMENTS"] = args.str();
-    stringstream load1;
-    for (int i = 0; i < paramBuffers.size(); i++)
-        load1<<paramBuffers[i].getType()<<" params"<<(i+1)<<"1 = global_params"<<(i+1)<<"[atom1];\n";
-    for (int i = 0; i < computedValueBuffers.size(); i++)
-        load1<<computedValueBuffers[i].getType()<<" values"<<(i+1)<<"1 = global_values"<<(i+1)<<"[atom1];\n";
-    replacements["LOAD_ATOM1_PARAMETERS"] = load1.str();
-    stringstream loadLocal2;
-    for (int i = 0; i < paramBuffers.size(); i++)
-        loadLocal2<<"localData[LOCAL_ID].params"<<(i+1)<<" = global_params"<<(i+1)<<"[atom2];\n";
-    for (int i = 0; i < computedValueBuffers.size(); i++)
-        loadLocal2<<"localData[LOCAL_ID].values"<<(i+1)<<" = global_values"<<(i+1)<<"[atom2];\n";
-    replacements["LOAD_LOCAL_PARAMETERS"] = loadLocal2.str();
-    stringstream load2;
-    for (int i = 0; i < paramBuffers.size(); i++)
-        load2<<paramBuffers[i].getType()<<" params"<<(i+1)<<"2 = localData[localIndex].params"<<(i+1)<<";\n";
-    for (int i = 0; i < computedValueBuffers.size(); i++)
-        load2<<computedValueBuffers[i].getType()<<" values"<<(i+1)<<"2 = localData[localIndex].values"<<(i+1)<<";\n";
-    replacements["LOAD_ATOM2_PARAMETERS"] = load2.str();
-    stringstream initDerivs, saveDerivs;
-    const vector<string>& allParamDerivNames = cc.getEnergyParamDerivNames();
-    int numDerivs = allParamDerivNames.size();
-    for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-        string paramName = force.getEnergyParameterDerivativeName(i);
-        string derivVariable = cc.getNonbondedUtilities().addEnergyParameterDerivative(paramName);
-        initDerivs<<"mixed "<<derivVariable<<" = 0;\n";
-        for (int index = 0; index < numDerivs; index++)
-            if (allParamDerivNames[index] == paramName)
-                saveDerivs<<"energyParamDerivs[GLOBAL_ID*numDerivatives+"<<index<<"] += "<<derivVariable<<";\n";
-    }
-    replacements["INIT_DERIVATIVES"] = initDerivs.str();
-    replacements["SAVE_DERIVATIVES"] = saveDerivs.str();
-    map<string, string> defines;
-    if (force.getNonbondedMethod() != CustomNonbondedForce::NoCutoff)
-        defines["USE_CUTOFF"] = "1";
-    if (force.getNonbondedMethod() == CustomNonbondedForce::CutoffPeriodic)
-        defines["USE_PERIODIC"] = "1";
-    int localMemorySize = max(32, cc.getNonbondedUtilities().getForceThreadBlockSize());
-    defines["LOCAL_MEMORY_SIZE"] = cc.intToString(localMemorySize);
-    defines["WARPS_IN_BLOCK"] = cc.intToString(localMemorySize/32);
-    double cutoff = force.getCutoffDistance();
-    defines["CUTOFF_SQUARED"] = cc.doubleToString(cutoff*cutoff);
-    double paddedCutoff = cc.getNonbondedUtilities().padCutoff(cutoff);
-    defines["PADDED_CUTOFF_SQUARED"] = cc.doubleToString(paddedCutoff*paddedCutoff);
-    defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-    defines["TILE_SIZE"] = "32";
-    defines["NUM_TILES"] = cc.intToString(numTileSets);
-    int numContexts = cc.getNumContexts();
-    int startIndex = cc.getContextIndex()*numTileSets/numContexts;
-    int endIndex = (cc.getContextIndex()+1)*numTileSets/numContexts;
-    defines["FIRST_TILE"] = cc.intToString(startIndex);
-    defines["LAST_TILE"] = cc.intToString(endIndex);
-    if ((localDataSize/4)%2 == 0 && !cc.getUseDoublePrecision())
-        defines["PARAMETER_SIZE_IS_EVEN"] = "1";
-    ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customNonbondedGroups, replacements), defines);
-    interactionGroupKernel = program->createKernel("computeInteractionGroups");
-    prepareNeighborListKernel = program->createKernel("prepareToBuildNeighborList");
-    buildNeighborListKernel = program->createKernel("buildNeighborList");
-    numGroupThreadBlocks = cc.getNonbondedUtilities().getNumForceThreadBlocks();
-}
-
-double CommonCalcCustomNonbondedForceKernel::execute(ContextImpl& context, bool includeForces, bool includeEnergy) {
-    useNeighborList = (filteredGroupData.isInitialized() && cc.getNonbondedUtilities().getUseCutoff());
-    if (useNeighborList && cc.getContextIndex() > 0) {
-        // When using a neighbor list, run the whole calculation on a single device.
-        return 0.0;
-    }
-    ContextSelector selector(cc);
-    bool recomputeLongRangeCorrection = !hasInitializedLongRangeCorrection;
-    if (globals.isInitialized()) {
-        bool changed = false;
-        for (int i = 0; i < (int) globalParamNames.size(); i++) {
-            float value = (float) context.getParameter(globalParamNames[i]);
-            if (value != globalParamValues[i])
-                changed = true;
-            globalParamValues[i] = value;
-        }
-        if (changed) {
-            globals.upload(globalParamValues);
-            if (forceCopy != NULL)
-                recomputeLongRangeCorrection = true;
-        }
-    }
-    if (recomputeLongRangeCorrection) {
-        if (includeEnergy || forceCopy->getNumEnergyParameterDerivatives() > 0) {
-            cc.getWorkThread().addTask(new LongRangeTask(cc, context.getOwner(), longRangeCorrectionData, longRangeCoefficient, longRangeCoefficientDerivs, forceCopy));
-            hasInitializedLongRangeCorrection = true;
-        }
-        else
-            hasInitializedLongRangeCorrection = false;
-    }
-    if (computedValues != NULL)
-        computedValuesKernel->execute(cc.getNumAtoms());
-    if (interactionGroupData.isInitialized()) {
-        if (!hasInitializedKernel) {
-            hasInitializedKernel = true;
-            interactionGroupKernel->addArg(cc.getLongForceBuffer());
-            interactionGroupKernel->addArg(cc.getEnergyBuffer());
-            interactionGroupKernel->addArg(cc.getPosq());
-            interactionGroupKernel->addArg((useNeighborList ? filteredGroupData : interactionGroupData));
-            interactionGroupKernel->addArg(numGroupTiles);
-            interactionGroupKernel->addArg((int) useNeighborList);
-            for (int i = 0; i < 5; i++)
-                interactionGroupKernel->addArg(); // Periodic box information will be set just before it is executed.
-            interactionGroupKernel->addArg((int) cc.getEnergyParamDerivNames().size());
-            for (auto& buffer : paramBuffers)
-                interactionGroupKernel->addArg(buffer.getArray());
-            for (auto& buffer : computedValueBuffers)
-                interactionGroupKernel->addArg(buffer.getArray());
-            for (auto& function : tabulatedFunctionArrays)
-                interactionGroupKernel->addArg(function);
-            if (globals.isInitialized())
-                interactionGroupKernel->addArg(globals);
-            if (hasParamDerivs)
-                interactionGroupKernel->addArg(cc.getEnergyParamDerivBuffer());
-            if (useNeighborList) {
-                // Initialize kernels for building the interaction group neighbor list.
-
-                prepareNeighborListKernel->addArg(cc.getNonbondedUtilities().getRebuildNeighborList());
-                prepareNeighborListKernel->addArg(numGroupTiles);
-                buildNeighborListKernel->addArg(cc.getNonbondedUtilities().getRebuildNeighborList());
-                buildNeighborListKernel->addArg(numGroupTiles);
-                buildNeighborListKernel->addArg(cc.getPosq());
-                buildNeighborListKernel->addArg(interactionGroupData);
-                buildNeighborListKernel->addArg(filteredGroupData);
-                for (int i = 0; i < 5; i++)
-                    buildNeighborListKernel->addArg(); // Periodic box information will be set just before it is executed.
-            }
-        }
-        int forceThreadBlockSize = max(32, cc.getNonbondedUtilities().getForceThreadBlockSize());
-        if (useNeighborList) {
-            // Rebuild the neighbor list, if necessary.
-
-            setPeriodicBoxArgs(cc, buildNeighborListKernel, 5);
-            prepareNeighborListKernel->execute(1, 1);
-            buildNeighborListKernel->execute(numGroupThreadBlocks*forceThreadBlockSize, forceThreadBlockSize);
-        }
-        setPeriodicBoxArgs(cc, interactionGroupKernel, 6);
-        interactionGroupKernel->execute(numGroupThreadBlocks*forceThreadBlockSize, forceThreadBlockSize);
-    }
-    return 0;
-}
-
-void CommonCalcCustomNonbondedForceKernel::copyParametersToContext(ContextImpl& context, const CustomNonbondedForce& force, int firstParticle, int lastParticle) {
-    ContextSelector selector(cc);
-    int numParticles = force.getNumParticles();
-    if (numParticles != cc.getNumAtoms())
-        throw OpenMMException("updateParametersInContext: The number of particles has changed");
-
-    // Record the per-particle parameters.
-
-    if (firstParticle <= lastParticle) {
-        int numToSet = lastParticle-firstParticle+1;
-        int numParams = force.getNumPerParticleParameters();
-        vector<vector<float> > paramVector(numToSet, vector<float>(numParams, 0));
-        vector<double> parameters;
-        for (int i = 0; i < numToSet; i++) {
-            force.getParticleParameters(firstParticle+i, parameters);
-            paramVector[i].resize(parameters.size());
-            for (int j = 0; j < (int) parameters.size(); j++)
-                paramVector[i][j] = (float) parameters[j];
-        }
-        params->setParameterValuesSubset(firstParticle, paramVector);
-    }
-
-    // If necessary, recompute the long range correction.
-
-    if (forceCopy != NULL) {
-        longRangeCorrectionData = CustomNonbondedForceImpl::prepareLongRangeCorrection(force, cc.getThreadPool().getNumThreads());
-        CustomNonbondedForceImpl::calcLongRangeCorrection(force, longRangeCorrectionData, context.getOwner(), longRangeCoefficient, longRangeCoefficientDerivs, cc.getThreadPool());
-        hasInitializedLongRangeCorrection = false;
-        *forceCopy = force;
-    }
-
-    // See if any tabulated functions have changed.
-
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        string name = force.getTabulatedFunctionName(i);
-        if (force.getTabulatedFunction(i).getUpdateCount() != tabulatedFunctionUpdateCount[name]) {
-            tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-            int width;
-            vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-            tabulatedFunctionArrays[i].upload(f);
-        }
-    }
-
-    // Mark that the current reordering may be invalid.
-
-    cc.invalidateMolecules(info, firstParticle <= lastParticle, false);
-}
-
 class CommonCalcGBSAOBCForceKernel::ForceInfo : public ComputeForceInfo {
 public:
     ForceInfo(const GBSAOBCForce& force) : force(force) {
@@ -2975,1025 +2205,6 @@ void CommonCalcGBSAOBCForceKernel::copyParametersToContext(ContextImpl& context,
     
     // Mark that the current reordering may be invalid.
     
-    cc.invalidateMolecules(info, true, false);
-}
-
-class CommonCalcCustomGBForceKernel::ForceInfo : public ComputeForceInfo {
-public:
-    ForceInfo(const CustomGBForce& force) : force(force) {
-    }
-    bool areParticlesIdentical(int particle1, int particle2) {
-        thread_local static vector<double> params1, params2;
-        force.getParticleParameters(particle1, params1);
-        force.getParticleParameters(particle2, params2);
-        for (int i = 0; i < (int) params1.size(); i++)
-            if (params1[i] != params2[i])
-                return false;
-        return true;
-    }
-    int getNumParticleGroups() {
-        return force.getNumExclusions();
-    }
-    void getParticlesInGroup(int index, vector<int>& particles) {
-        int particle1, particle2;
-        force.getExclusionParticles(index, particle1, particle2);
-        particles.resize(2);
-        particles[0] = particle1;
-        particles[1] = particle2;
-    }
-    bool areGroupsIdentical(int group1, int group2) {
-        return true;
-    }
-private:
-    const CustomGBForce& force;
-};
-
-CommonCalcCustomGBForceKernel::~CommonCalcCustomGBForceKernel() {
-    ContextSelector selector(cc);
-    if (params != NULL)
-        delete params;
-    if (computedValues != NULL)
-        delete computedValues;
-    if (energyDerivs != NULL)
-        delete energyDerivs;
-    if (energyDerivChain != NULL)
-        delete energyDerivChain;
-    for (auto d : dValuedParam)
-        delete d;
-}
-
-void CommonCalcCustomGBForceKernel::initialize(const System& system, const CustomGBForce& force) {
-    ContextSelector selector(cc);
-    if (cc.getNumContexts() > 1)
-        throw OpenMMException("CustomGBForce does not support using multiple devices");
-    NonbondedUtilities& nb = cc.getNonbondedUtilities();
-    cutoff = force.getCutoffDistance();
-    bool useExclusionsForValue = false;
-    numComputedValues = force.getNumComputedValues();
-    vector<string> computedValueNames(numComputedValues);
-    vector<string> computedValueExpressions(numComputedValues);
-    if (numComputedValues > 0) {
-        CustomGBForce::ComputationType type;
-        force.getComputedValueParameters(0, computedValueNames[0], computedValueExpressions[0], type);
-        if (type == CustomGBForce::SingleParticle)
-            throw OpenMMException("The first computed value for a CustomGBForce must be of type ParticlePair or ParticlePairNoExclusions.");
-        useExclusionsForValue = (type == CustomGBForce::ParticlePair);
-        for (int i = 1; i < numComputedValues; i++) {
-            force.getComputedValueParameters(i, computedValueNames[i], computedValueExpressions[i], type);
-            if (type != CustomGBForce::SingleParticle)
-                throw OpenMMException("A CustomGBForce may only have one computed value of type ParticlePair or ParticlePairNoExclusions.");
-        }
-    }
-    int forceIndex;
-    for (forceIndex = 0; forceIndex < system.getNumForces() && &system.getForce(forceIndex) != &force; ++forceIndex)
-        ;
-    string prefix = "custom"+cc.intToString(forceIndex)+"_";
-
-    // Record parameters and exclusions.
-
-    int numParticles = force.getNumParticles();
-    int paddedNumParticles = cc.getPaddedNumAtoms();
-    int numParams = force.getNumPerParticleParameters();
-    params = new ComputeParameterSet(cc, force.getNumPerParticleParameters(), paddedNumParticles, "customGBParameters", true);
-    computedValues = new ComputeParameterSet(cc, numComputedValues, paddedNumParticles, "customGBComputedValues", true, cc.getUseDoublePrecision());
-    if (force.getNumGlobalParameters() > 0)
-        globals.initialize<float>(cc, force.getNumGlobalParameters(), "customGBGlobals");
-    vector<vector<float> > paramVector(paddedNumParticles, vector<float>(numParams, 0));
-    vector<vector<int> > exclusionList(numParticles);
-    for (int i = 0; i < numParticles; i++) {
-        vector<double> parameters;
-        force.getParticleParameters(i, parameters);
-        for (int j = 0; j < (int) parameters.size(); j++)
-            paramVector[i][j] = (float) parameters[j];
-        exclusionList[i].push_back(i);
-    }
-    for (int i = 0; i < force.getNumExclusions(); i++) {
-        int particle1, particle2;
-        force.getExclusionParticles(i, particle1, particle2);
-        exclusionList[particle1].push_back(particle2);
-        exclusionList[particle2].push_back(particle1);
-    }
-    params->setParameterValues(paramVector);
-
-    // Record the tabulated functions.
-
-    map<string, Lepton::CustomFunction*> functions;
-    vector<pair<string, string> > functionDefinitions;
-    vector<const TabulatedFunction*> functionList;
-    stringstream tableArgs;
-    tabulatedFunctionArrays.resize(force.getNumTabulatedFunctions());
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        functionList.push_back(&force.getTabulatedFunction(i));
-        string name = force.getTabulatedFunctionName(i);
-        tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-        string arrayName = prefix+"table"+cc.intToString(i);
-        functionDefinitions.push_back(make_pair(name, arrayName));
-        functions[name] = cc.getExpressionUtilities().getFunctionPlaceholder(force.getTabulatedFunction(i));
-        int width;
-        vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-        tabulatedFunctionArrays[i].initialize<float>(cc, f.size(), "TabulatedFunction");
-        tabulatedFunctionArrays[i].upload(f);
-        nb.addArgument(ComputeParameterInfo(tabulatedFunctionArrays[i], arrayName, "float", width));
-        tableArgs << ", GLOBAL const float";
-        if (width > 1)
-            tableArgs << width;
-        tableArgs << "* RESTRICT " << arrayName;
-    }
-
-    // Record the global parameters.
-
-    globalParamNames.resize(force.getNumGlobalParameters());
-    globalParamValues.resize(force.getNumGlobalParameters());
-    for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-        globalParamNames[i] = force.getGlobalParameterName(i);
-        globalParamValues[i] = (float) force.getGlobalParameterDefaultValue(i);
-    }
-    if (globals.isInitialized())
-        globals.upload(globalParamValues);
-
-    // Record derivatives of expressions needed for the chain rule terms.
-
-    vector<vector<Lepton::ParsedExpression> > valueGradientExpressions(numComputedValues);
-    vector<vector<Lepton::ParsedExpression> > valueDerivExpressions(numComputedValues);
-    vector<vector<Lepton::ParsedExpression> > valueParamDerivExpressions(numComputedValues);
-    needParameterGradient = false;
-    for (int i = 0; i < numComputedValues; i++) {
-        Lepton::ParsedExpression ex = Lepton::Parser::parse(computedValueExpressions[i], functions).optimize();
-        if (i > 0) {
-            valueGradientExpressions[i].push_back(ex.differentiate("x").optimize());
-            valueGradientExpressions[i].push_back(ex.differentiate("y").optimize());
-            valueGradientExpressions[i].push_back(ex.differentiate("z").optimize());
-            if (!isZeroExpression(valueGradientExpressions[i][0]) || !isZeroExpression(valueGradientExpressions[i][1]) || !isZeroExpression(valueGradientExpressions[i][2]))
-                needParameterGradient = true;
-            for (int j = 0; j < i; j++)
-                valueDerivExpressions[i].push_back(ex.differentiate(computedValueNames[j]).optimize());
-        }
-        for (int j = 0; j < force.getNumEnergyParameterDerivatives(); j++)
-            valueParamDerivExpressions[i].push_back(ex.differentiate(force.getEnergyParameterDerivativeName(j)).optimize());
-    }
-    vector<vector<Lepton::ParsedExpression> > energyDerivExpressions(force.getNumEnergyTerms());
-    vector<vector<Lepton::ParsedExpression> > energyParamDerivExpressions(force.getNumEnergyTerms());
-    vector<bool> needChainForValue(numComputedValues, false);
-    for (int i = 0; i < force.getNumEnergyTerms(); i++) {
-        string expression;
-        CustomGBForce::ComputationType type;
-        force.getEnergyTermParameters(i, expression, type);
-        Lepton::ParsedExpression ex = Lepton::Parser::parse(expression, functions).optimize();
-        for (int j = 0; j < numComputedValues; j++) {
-            if (type == CustomGBForce::SingleParticle) {
-                energyDerivExpressions[i].push_back(ex.differentiate(computedValueNames[j]).optimize());
-                if (!isZeroExpression(energyDerivExpressions[i].back()))
-                    needChainForValue[j] = true;
-            }
-            else {
-                energyDerivExpressions[i].push_back(ex.differentiate(computedValueNames[j]+"1").optimize());
-                if (!isZeroExpression(energyDerivExpressions[i].back()))
-                    needChainForValue[j] = true;
-                energyDerivExpressions[i].push_back(ex.differentiate(computedValueNames[j]+"2").optimize());
-                if (!isZeroExpression(energyDerivExpressions[i].back()))
-                    needChainForValue[j] = true;
-            }
-        }
-        for (int j = 0; j < force.getNumEnergyParameterDerivatives(); j++)
-            energyParamDerivExpressions[i].push_back(ex.differentiate(force.getEnergyParameterDerivativeName(j)).optimize());
-    }
-    bool deviceIsCpu = cc.getIsCPU();
-    int elementSize = (cc.getUseDoublePrecision() ? sizeof(double) : sizeof(float));
-    valueBuffers.initialize<long long>(cc, cc.getPaddedNumAtoms(), "customGBValueBuffers");
-    longEnergyDerivs.initialize<long long>(cc, numComputedValues*cc.getPaddedNumAtoms(), "customGBLongEnergyDerivatives");
-    energyDerivs = new ComputeParameterSet(cc, numComputedValues, cc.getPaddedNumAtoms(), "customGBEnergyDerivatives", true);
-    cc.addAutoclearBuffer(valueBuffers);
-    energyDerivChain = new ComputeParameterSet(cc, numComputedValues, cc.getPaddedNumAtoms(), "customGBEnergyDerivativeChain", true);
-    needEnergyParamDerivs = (force.getNumEnergyParameterDerivatives() > 0);
-    dValue0dParam.resize(force.getNumEnergyParameterDerivatives());
-    for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-        dValuedParam.push_back(new ComputeParameterSet(cc, numComputedValues, cc.getPaddedNumAtoms(), "dValuedParam", true, cc.getUseDoublePrecision()));
-        dValue0dParam[i].initialize<long long>(cc, cc.getPaddedNumAtoms(), "dValue0dParam");
-        cc.addAutoclearBuffer(dValue0dParam[i]);
-        string name = force.getEnergyParameterDerivativeName(i);
-        cc.addEnergyParameterDerivative(name);
-    }
-
-    // Create the kernels.
-
-    bool useCutoff = (force.getNonbondedMethod() != CustomGBForce::NoCutoff);
-    bool usePeriodic = (force.getNonbondedMethod() != CustomGBForce::NoCutoff && force.getNonbondedMethod() != CustomGBForce::CutoffNonPeriodic);
-    int numAtomBlocks = cc.getPaddedNumAtoms()/32;
-    {
-        // Create the N2 value kernel.
-
-        vector<pair<ExpressionTreeNode, string> > variables;
-        map<string, string> rename;
-        ExpressionTreeNode rnode(new Operation::Variable("r"));
-        variables.push_back(make_pair(rnode, "r"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Square(), rnode), "r2"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Reciprocal(), rnode), "invR"));
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++) {
-            const string& name = force.getPerParticleParameterName(i);
-            variables.push_back(makeVariable(name+"1", "((real) params"+params->getParameterSuffix(i, "1)")));
-            variables.push_back(makeVariable(name+"2", "((real) params"+params->getParameterSuffix(i, "2)")));
-            rename[name+"1"] = name+"2";
-            rename[name+"2"] = name+"1";
-        }
-        for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-            const string& name = force.getGlobalParameterName(i);
-            string value = "globals["+cc.intToString(i)+"]";
-            variables.push_back(makeVariable(name, value));
-        }
-        map<string, Lepton::ParsedExpression> n2ValueExpressions;
-        stringstream n2ValueSource;
-        Lepton::ParsedExpression ex = Lepton::Parser::parse(computedValueExpressions[0], functions).optimize();
-        n2ValueExpressions["tempValue1 = "] = ex;
-        n2ValueExpressions["tempValue2 = "] = ex.renameVariables(rename);
-        for (int i = 0; i < valueParamDerivExpressions[0].size(); i++) {
-            string variableBase = "temp_dValue0dParam"+cc.intToString(i+1);
-            if (!isZeroExpression(valueParamDerivExpressions[0][i])) {
-                n2ValueExpressions[variableBase+"_1 = "] = valueParamDerivExpressions[0][i];
-                n2ValueExpressions[variableBase+"_2 = "] = valueParamDerivExpressions[0][i].renameVariables(rename);
-            }
-        }
-        n2ValueSource << cc.getExpressionUtilities().createExpressions(n2ValueExpressions, variables, functionList, functionDefinitions, "temp");
-        map<string, string> replacements;
-        string n2ValueStr = n2ValueSource.str();
-        replacements["COMPUTE_VALUE"] = n2ValueStr;
-        stringstream extraArgs, atomParams, loadLocal1, loadLocal2, load1, load2, tempDerivs1, tempDerivs2, storeDeriv1, storeDeriv2;
-        if (force.getNumGlobalParameters() > 0)
-            extraArgs << ", GLOBAL const float* globals";
-        pairValueUsesParam.resize(params->getParameterInfos().size(), false);
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            if (n2ValueStr.find(paramName+"1") != n2ValueStr.npos || n2ValueStr.find(paramName+"2") != n2ValueStr.npos) {
-                extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT global_" << paramName;
-                atomParams << "LOCAL " << buffer.getType() << " local_" << paramName << "[LOCAL_BUFFER_SIZE];\n";
-                loadLocal1 << "local_" << paramName << "[localAtomIndex] = " << paramName << "1;\n";
-                loadLocal2 << "local_" << paramName << "[localAtomIndex] = global_" << paramName << "[j];\n";
-                load1 << buffer.getType() << " " << paramName << "1 = global_" << paramName << "[atom1];\n";
-                load2 << buffer.getType() << " " << paramName << "2 = local_" << paramName << "[atom2];\n";
-                pairValueUsesParam[i] = true;
-            }
-        }
-        for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-            string derivName = "dValue0dParam"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL mm_ulong* RESTRICT global_" << derivName;
-            atomParams << "LOCAL real local_" << derivName << "[LOCAL_BUFFER_SIZE];\n";
-            loadLocal2 << "local_" << derivName << "[localAtomIndex] = 0;\n";
-            load1 << "real " << derivName << " = 0;\n";
-            if (!isZeroExpression(valueParamDerivExpressions[0][i])) {
-                load2 << "real temp_" << derivName << "_1 = 0;\n";
-                load2 << "real temp_" << derivName << "_2 = 0;\n";
-                tempDerivs1 << derivName << " += temp_" << derivName << "_1;\n";
-                if (deviceIsCpu)
-                    tempDerivs2 << "local_" << derivName << "[j] += temp_" << derivName << "_2;\n";
-                else
-                    tempDerivs2 << "local_" << derivName << "[tbx+tj] += temp_" << derivName << "_2;\n";
-                storeDeriv1 << "ATOMIC_ADD(&global_" << derivName << "[offset1], (mm_ulong) realToFixedPoint(" << derivName << "));\n";
-                if (deviceIsCpu)
-                    storeDeriv2 << "ATOMIC_ADD(&global_" << derivName << "[offset2], (mm_ulong) realToFixedPoint(local_" << derivName << "[tgx]));\n";
-                else
-                    storeDeriv2 << "ATOMIC_ADD(&global_" << derivName << "[offset2], (mm_ulong) realToFixedPoint(local_" << derivName << "[LOCAL_ID]));\n";
-            }
-        }
-        replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-        replacements["ATOM_PARAMETER_DATA"] = atomParams.str();
-        replacements["LOAD_LOCAL_PARAMETERS_FROM_1"] = loadLocal1.str();
-        replacements["LOAD_LOCAL_PARAMETERS_FROM_GLOBAL"] = loadLocal2.str();
-        replacements["LOAD_ATOM1_PARAMETERS"] = load1.str();
-        replacements["LOAD_ATOM2_PARAMETERS"] = load2.str();
-        replacements["ADD_TEMP_DERIVS1"] = tempDerivs1.str();
-        replacements["ADD_TEMP_DERIVS2"] = tempDerivs2.str();
-        replacements["STORE_PARAM_DERIVS1"] = storeDeriv1.str();
-        replacements["STORE_PARAM_DERIVS2"] = storeDeriv2.str();
-        if (useCutoff)
-            pairValueDefines["USE_CUTOFF"] = "1";
-        if (usePeriodic)
-            pairValueDefines["USE_PERIODIC"] = "1";
-        if (useExclusionsForValue)
-            pairValueDefines["USE_EXCLUSIONS"] = "1";
-        pairValueDefines["LOCAL_BUFFER_SIZE"] = cc.intToString(deviceIsCpu ? 32 : nb.getForceThreadBlockSize());
-        pairValueDefines["CUTOFF_SQUARED"] = cc.doubleToString(cutoff*cutoff);
-        pairValueDefines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        pairValueDefines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-        pairValueDefines["NUM_BLOCKS"] = cc.intToString(numAtomBlocks);
-        pairValueDefines["TILE_SIZE"] = "32";
-        string file;
-        if (deviceIsCpu)
-            file = CommonKernelSources::customGBValueN2_cpu;
-        else
-            file = CommonKernelSources::customGBValueN2;
-        pairValueSrc = cc.replaceStrings(file, replacements);
-    }
-    {
-        // Create the kernel to reduce the N2 value and calculate other values.
-
-        stringstream reductionSource, extraArgs, deriv0;
-        if (force.getNumGlobalParameters() > 0)
-            extraArgs << ", GLOBAL const float* globals";
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << paramName;
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string valueName = "values"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL " << buffer.getType() << "* RESTRICT global_" << valueName;
-            reductionSource << buffer.getType() << " local_" << valueName << ";\n";
-        }
-        for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-            string variableName = "dValuedParam_0_"+cc.intToString(i);
-            extraArgs << ", GLOBAL const mm_long* RESTRICT dValue0dParam" << i;
-            deriv0 << "real " << variableName << " = RECIP((real) 0x100000000)*dValue0dParam" << i << "[index];\n";
-            for (int j = 0; j < dValuedParam[i]->getParameterInfos().size(); j++)
-                extraArgs << ", GLOBAL real* RESTRICT global_dValuedParam_" << j << "_" << i;
-            deriv0 << "global_dValuedParam_0_" << i << "[index] = dValuedParam_0_" << i << ";\n";
-        }
-        reductionSource << "local_values" << computedValues->getParameterSuffix(0) << " = sum;\n";
-        map<string, string> variables;
-        variables["x"] = "pos.x";
-        variables["y"] = "pos.y";
-        variables["z"] = "pos.z";
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++)
-            variables[force.getPerParticleParameterName(i)] = "params"+params->getParameterSuffix(i, "[index]");
-        for (int i = 0; i < force.getNumGlobalParameters(); i++)
-            variables[force.getGlobalParameterName(i)] = "globals["+cc.intToString(i)+"]";
-        for (int i = 1; i < numComputedValues; i++) {
-            variables[computedValueNames[i-1]] = "local_values"+computedValues->getParameterSuffix(i-1);
-            map<string, Lepton::ParsedExpression> valueExpressions;
-            valueExpressions["local_values"+computedValues->getParameterSuffix(i)+" = "] = Lepton::Parser::parse(computedValueExpressions[i], functions).optimize();
-            reductionSource << cc.getExpressionUtilities().createExpressions(valueExpressions, variables, functionList, functionDefinitions, "value"+cc.intToString(i)+"_temp");
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            string valueName = "values"+cc.intToString(i+1);
-            reductionSource << "global_" << valueName << "[index] = local_" << valueName << ";\n";
-        }
-        if (needEnergyParamDerivs) {
-            map<string, Lepton::ParsedExpression> derivExpressions;
-            for (int i = 1; i < numComputedValues; i++) {
-                for (int j = 0; j < valueParamDerivExpressions[i].size(); j++)
-                    derivExpressions["real dValuedParam_"+cc.intToString(i)+"_"+cc.intToString(j)+" = "] = valueParamDerivExpressions[i][j];
-                for (int j = 0; j < i; j++)
-                    derivExpressions["real dVdV_"+cc.intToString(i)+"_"+cc.intToString(j)+" = "] = valueDerivExpressions[i][j];
-            }
-            reductionSource << cc.getExpressionUtilities().createExpressions(derivExpressions, variables, functionList, functionDefinitions, "derivChain_temp");
-            for (int i = 1; i < numComputedValues; i++) {
-                for (int j = 0; j < i; j++)
-                    for (int k = 0; k < valueParamDerivExpressions[i].size(); k++)
-                        reductionSource << "dValuedParam_" << i << "_" << k << " += dVdV_" << i << "_" << j << "*dValuedParam_" << j <<"_" << k << ";\n";
-                for (int j = 0; j < valueParamDerivExpressions[i].size(); j++)
-                    reductionSource << "global_dValuedParam_" << i << "_" << j << "[index] = dValuedParam_" << i << "_" << j << ";\n";
-            }
-        }
-        map<string, string> replacements;
-        replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-        replacements["REDUCE_PARAM0_DERIV"] = deriv0.str();
-        replacements["COMPUTE_VALUES"] = reductionSource.str();
-        map<string, string> defines;
-        defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customGBValuePerParticle, replacements), defines);
-        perParticleValueKernel = program->createKernel("computePerParticleValues");
-    }
-    {
-        // Create the N2 energy kernel.
-
-        vector<pair<ExpressionTreeNode, string> > variables;
-        ExpressionTreeNode rnode(new Operation::Variable("r"));
-        variables.push_back(make_pair(rnode, "r"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Square(), rnode), "r2"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Reciprocal(), rnode), "invR"));
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++) {
-            const string& name = force.getPerParticleParameterName(i);
-            variables.push_back(makeVariable(name+"1", "((real) params"+params->getParameterSuffix(i, "1)")));
-            variables.push_back(makeVariable(name+"2", "((real) params"+params->getParameterSuffix(i, "2)")));
-        }
-        for (int i = 0; i < numComputedValues; i++) {
-            variables.push_back(makeVariable(computedValueNames[i]+"1", "values"+computedValues->getParameterSuffix(i, "1")));
-            variables.push_back(makeVariable(computedValueNames[i]+"2", "values"+computedValues->getParameterSuffix(i, "2")));
-        }
-        for (int i = 0; i < force.getNumGlobalParameters(); i++)
-            variables.push_back(makeVariable(force.getGlobalParameterName(i), "globals["+cc.intToString(i)+"]"));
-        stringstream n2EnergySource;
-        bool anyExclusions = (force.getNumExclusions() > 0);
-        for (int i = 0; i < force.getNumEnergyTerms(); i++) {
-            string expression;
-            CustomGBForce::ComputationType type;
-            force.getEnergyTermParameters(i, expression, type);
-            if (type == CustomGBForce::SingleParticle)
-                continue;
-            bool exclude = (anyExclusions && type == CustomGBForce::ParticlePair);
-            map<string, Lepton::ParsedExpression> n2EnergyExpressions;
-            n2EnergyExpressions["tempEnergy += "] = Lepton::Parser::parse(expression, functions).optimize();
-            n2EnergyExpressions["dEdR += "] = Lepton::Parser::parse(expression, functions).differentiate("r").optimize();
-            for (int j = 0; j < numComputedValues; j++) {
-                if (needChainForValue[j]) {
-                    string index = cc.intToString(j+1);
-                    n2EnergyExpressions["/*"+cc.intToString(i+1)+"*/ deriv"+index+"_1 += "] = energyDerivExpressions[i][2*j];
-                    n2EnergyExpressions["/*"+cc.intToString(i+1)+"*/ deriv"+index+"_2 += "] = energyDerivExpressions[i][2*j+1];
-                }
-            }
-            for (int j = 0; j < force.getNumEnergyParameterDerivatives(); j++)
-                n2EnergyExpressions["energyParamDeriv"+cc.intToString(j)+" += interactionScale*"] = energyParamDerivExpressions[i][j];
-            if (exclude)
-                n2EnergySource << "if (!isExcluded) {\n";
-            n2EnergySource << cc.getExpressionUtilities().createExpressions(n2EnergyExpressions, variables, functionList, functionDefinitions, "temp");
-            if (exclude)
-                n2EnergySource << "}\n";
-        }
-        map<string, string> replacements;
-        string n2EnergyStr = n2EnergySource.str();
-        replacements["COMPUTE_INTERACTION"] = n2EnergyStr;
-        stringstream extraArgs, atomParams, loadLocal1, loadLocal2, clearLocal, load1, load2, declare1, recordDeriv, storeDerivs1, storeDerivs2, initParamDerivs, saveParamDerivs;
-        if (force.getNumGlobalParameters() > 0)
-            extraArgs << ", GLOBAL const float* globals";
-        pairEnergyUsesParam.resize(params->getParameterInfos().size(), false);
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            if (n2EnergyStr.find(paramName+"1") != n2EnergyStr.npos || n2EnergyStr.find(paramName+"2") != n2EnergyStr.npos) {
-                extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT global_" << paramName;
-                atomParams << "LOCAL " << buffer.getType() << " local_" << paramName << "[LOCAL_BUFFER_SIZE];\n";
-                loadLocal1 << "local_" << paramName << "[localAtomIndex] = " << paramName << "1;\n";
-                loadLocal2 << "local_" << paramName << "[localAtomIndex] = global_" << paramName << "[j];\n";
-                load1 << buffer.getType() << " " << paramName << "1 = global_" << paramName << "[atom1];\n";
-                load2 << buffer.getType() << " " << paramName << "2 = local_" << paramName << "[atom2];\n";
-                pairEnergyUsesParam[i] = true;
-            }
-        }
-        pairEnergyUsesValue.resize(computedValues->getParameterInfos().size(), false);
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string valueName = "values"+cc.intToString(i+1);
-            if (n2EnergyStr.find(valueName+"1") != n2EnergyStr.npos || n2EnergyStr.find(valueName+"2") != n2EnergyStr.npos) {
-                extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT global_" << valueName;
-                atomParams << "LOCAL " << buffer.getType() << " local_" << valueName << "[LOCAL_BUFFER_SIZE];\n";
-                loadLocal1 << "local_" << valueName << "[localAtomIndex] = " << valueName << "1;\n";
-                loadLocal2 << "local_" << valueName << "[localAtomIndex] = global_" << valueName << "[j];\n";
-                load1 << buffer.getType() << " " << valueName << "1 = global_" << valueName << "[atom1];\n";
-                load2 << buffer.getType() << " " << valueName << "2 = local_" << valueName << "[atom2];\n";
-                pairEnergyUsesValue[i] = true;
-            }
-        }
-        extraArgs << ", GLOBAL mm_ulong* RESTRICT derivBuffers";
-        for (int i = 0; i < numComputedValues; i++) {
-            string index = cc.intToString(i+1);
-            atomParams << "LOCAL real local_deriv" << index << "[LOCAL_BUFFER_SIZE];\n";
-            clearLocal << "local_deriv" << index << "[localAtomIndex] = 0.0f;\n";
-            declare1 << "real deriv" << index << "_1 = 0;\n";
-            load2 << "real deriv" << index << "_2 = 0;\n";
-            recordDeriv << "local_deriv" << index << "[atom2] += deriv" << index << "_2;\n";
-            storeDerivs1 << "STORE_DERIVATIVE_1(" << index << ")\n";
-            storeDerivs2 << "STORE_DERIVATIVE_2(" << index << ")\n";
-        }
-        if (needEnergyParamDerivs) {
-            extraArgs << ", GLOBAL mixed* RESTRICT energyParamDerivs";
-            const vector<string>& allParamDerivNames = cc.getEnergyParamDerivNames();
-            int numDerivs = allParamDerivNames.size();
-            for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-                initParamDerivs << "mixed energyParamDeriv" << i << " = 0;\n";
-                for (int index = 0; index < numDerivs; index++)
-                    if (allParamDerivNames[index] == force.getEnergyParameterDerivativeName(i))
-                        saveParamDerivs << "energyParamDerivs[GLOBAL_ID*" << numDerivs << "+" << index << "] += energyParamDeriv" << i << ";\n";
-            }
-        }
-        replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-        replacements["ATOM_PARAMETER_DATA"] = atomParams.str();
-        replacements["LOAD_LOCAL_PARAMETERS_FROM_1"] = loadLocal1.str();
-        replacements["LOAD_LOCAL_PARAMETERS_FROM_GLOBAL"] = loadLocal2.str();
-        replacements["CLEAR_LOCAL_DERIVATIVES"] = clearLocal.str();
-        replacements["LOAD_ATOM1_PARAMETERS"] = load1.str();
-        replacements["LOAD_ATOM2_PARAMETERS"] = load2.str();
-        replacements["DECLARE_ATOM1_DERIVATIVES"] = declare1.str();
-        replacements["RECORD_DERIVATIVE_2"] = recordDeriv.str();
-        replacements["STORE_DERIVATIVES_1"] = storeDerivs1.str();
-        replacements["STORE_DERIVATIVES_2"] = storeDerivs2.str();
-        replacements["INIT_PARAM_DERIVS"] = initParamDerivs.str();
-        replacements["SAVE_PARAM_DERIVS"] = saveParamDerivs.str();
-        if (useCutoff)
-            pairEnergyDefines["USE_CUTOFF"] = "1";
-        if (usePeriodic)
-            pairEnergyDefines["USE_PERIODIC"] = "1";
-        if (anyExclusions)
-            pairEnergyDefines["USE_EXCLUSIONS"] = "1";
-        pairEnergyDefines["LOCAL_BUFFER_SIZE"] = cc.intToString(deviceIsCpu ? 32 : nb.getForceThreadBlockSize());
-        pairEnergyDefines["CUTOFF_SQUARED"] = cc.doubleToString(cutoff*cutoff);
-        pairEnergyDefines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        pairEnergyDefines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-        pairEnergyDefines["NUM_BLOCKS"] = cc.intToString(numAtomBlocks);
-        pairEnergyDefines["TILE_SIZE"] = "32";
-        string file;
-        if (deviceIsCpu)
-            file = CommonKernelSources::customGBEnergyN2_cpu;
-        else
-            file = CommonKernelSources::customGBEnergyN2;
-        pairEnergySrc = cc.replaceStrings(file, replacements);
-    }
-    {
-        // Create the kernel to reduce the derivatives and calculate per-particle energy terms.
-
-        stringstream compute, extraArgs, reduce, initParamDerivs, saveParamDerivs;
-        if (force.getNumGlobalParameters() > 0)
-            extraArgs << ", GLOBAL const float* globals";
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << paramName;
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string valueName = "values"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << valueName;
-        }
-        for (int i = 0; i < (int) energyDerivs->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = energyDerivs->getParameterInfos()[i];
-            string index = cc.intToString(i+1);
-            extraArgs << ", GLOBAL " << buffer.getType() << "* RESTRICT derivBuffers" << index;
-            compute << buffer.getType() << " deriv" << index << " = derivBuffers" << index << "[index];\n";
-        }
-        for (int i = 0; i < (int) energyDerivChain->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = energyDerivChain->getParameterInfos()[i];
-            string index = cc.intToString(i+1);
-            extraArgs << ", GLOBAL " << buffer.getType() << "* RESTRICT derivChain" << index;
-        }
-        extraArgs << ", GLOBAL const mm_long* RESTRICT derivBuffersIn";
-        for (int i = 0; i < energyDerivs->getNumParameters(); ++i)
-            reduce << "derivBuffers" << energyDerivs->getParameterSuffix(i, "[index]") <<
-                    " = RECIP((real) 0x100000000)*derivBuffersIn[index+PADDED_NUM_ATOMS*" << cc.intToString(i) << "];\n";
-        if (needEnergyParamDerivs) {
-            extraArgs << ", GLOBAL mixed* RESTRICT energyParamDerivs";
-            const vector<string>& allParamDerivNames = cc.getEnergyParamDerivNames();
-            int numDerivs = allParamDerivNames.size();
-            for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-                initParamDerivs << "mixed energyParamDeriv" << i << " = 0;\n";
-                for (int index = 0; index < numDerivs; index++)
-                    if (allParamDerivNames[index] == force.getEnergyParameterDerivativeName(i))
-                        saveParamDerivs << "energyParamDerivs[GLOBAL_ID*" << numDerivs << "+" << index << "] += energyParamDeriv" << i << ";\n";
-            }
-        }
-        
-        // Compute the various expressions.
-        
-        map<string, string> variables;
-        variables["x"] = "pos.x";
-        variables["y"] = "pos.y";
-        variables["z"] = "pos.z";
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++)
-            variables[force.getPerParticleParameterName(i)] = "params"+params->getParameterSuffix(i, "[index]");
-        for (int i = 0; i < force.getNumGlobalParameters(); i++)
-            variables[force.getGlobalParameterName(i)] = "globals["+cc.intToString(i)+"]";
-        for (int i = 0; i < numComputedValues; i++)
-            variables[computedValueNames[i]] = "values"+computedValues->getParameterSuffix(i, "[index]");
-        map<string, Lepton::ParsedExpression> expressions;
-        for (int i = 0; i < force.getNumEnergyTerms(); i++) {
-            string expression;
-            CustomGBForce::ComputationType type;
-            force.getEnergyTermParameters(i, expression, type);
-            if (type != CustomGBForce::SingleParticle)
-                continue;
-            Lepton::ParsedExpression parsed = Lepton::Parser::parse(expression, functions).optimize();
-            expressions["/*"+cc.intToString(i+1)+"*/ energy += "] = parsed;
-            for (int j = 0; j < numComputedValues; j++)
-                expressions["/*"+cc.intToString(i+1)+"*/ deriv"+energyDerivs->getParameterSuffix(j)+" += "] = energyDerivExpressions[i][j];
-            Lepton::ParsedExpression gradx = parsed.differentiate("x").optimize();
-            Lepton::ParsedExpression grady = parsed.differentiate("y").optimize();
-            Lepton::ParsedExpression gradz = parsed.differentiate("z").optimize();
-            if (!isZeroExpression(gradx))
-                expressions["/*"+cc.intToString(i+1)+"*/ force.x -= "] = gradx;
-            if (!isZeroExpression(grady))
-                expressions["/*"+cc.intToString(i+1)+"*/ force.y -= "] = grady;
-            if (!isZeroExpression(gradz))
-                expressions["/*"+cc.intToString(i+1)+"*/ force.z -= "] = gradz;
-            for (int j = 0; j < force.getNumEnergyParameterDerivatives(); j++)
-                expressions["/*"+cc.intToString(i+1)+"*/ energyParamDeriv"+cc.intToString(j)+" += "] = energyParamDerivExpressions[i][j];
-        }
-        for (int i = 1; i < numComputedValues; i++)
-            for (int j = 0; j < i; j++)
-                expressions["real dV"+cc.intToString(i)+"dV"+cc.intToString(j)+" = "] = valueDerivExpressions[i][j];
-        compute << cc.getExpressionUtilities().createExpressions(expressions, variables, functionList, functionDefinitions, "temp");
-        
-        // Record values.
-        
-        for (int i = 0; i < (int) energyDerivs->getParameterInfos().size(); i++) {
-            string index = cc.intToString(i+1);
-            compute << "derivBuffers" << index << "[index] = deriv" << index << ";\n";
-        }
-        compute << "forceBuffers[index] += realToFixedPoint(force.x);\n";
-        compute << "forceBuffers[index+PADDED_NUM_ATOMS] += realToFixedPoint(force.y);\n";
-        compute << "forceBuffers[index+PADDED_NUM_ATOMS*2] += realToFixedPoint(force.z);\n";
-        for (int i = 1; i < numComputedValues; i++) {
-            compute << "real totalDeriv"<<i<<" = dV"<<i<<"dV0";
-            for (int j = 1; j < i; j++)
-                compute << " + totalDeriv"<<j<<"*dV"<<i<<"dV"<<j;
-            compute << ";\n";
-            compute << "deriv"<<(i+1)<<" *= totalDeriv"<<i<<";\n";
-        }
-        for (int i = 0; i < (int) energyDerivs->getParameterInfos().size(); i++) {
-            string index = cc.intToString(i+1);
-            compute << "derivChain" << index << "[index] = deriv" << index << ";\n";
-        }
-        map<string, string> replacements;
-        replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-        replacements["REDUCE_DERIVATIVES"] = reduce.str();
-        replacements["COMPUTE_ENERGY"] = compute.str();
-        replacements["INIT_PARAM_DERIVS"] = initParamDerivs.str();
-        replacements["SAVE_PARAM_DERIVS"] = saveParamDerivs.str();
-        map<string, string> defines;
-        defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-        ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customGBEnergyPerParticle, replacements), defines);
-        perParticleEnergyKernel = program->createKernel("computePerParticleEnergy");
-    }
-    if (needParameterGradient || needEnergyParamDerivs) {
-        // Create the kernel to compute chain rule terms for computed values that depend explicitly on particle coordinates, and for
-        // derivatives with respect to global parameters.
-
-        stringstream compute, extraArgs, initParamDerivs, saveParamDerivs;
-        if (force.getNumGlobalParameters() > 0)
-            extraArgs << ", GLOBAL const float* globals";
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = "params"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << paramName;
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string valueName = "values"+cc.intToString(i+1);
-            extraArgs << ", GLOBAL const " << buffer.getType() << "* RESTRICT " << valueName;
-        }
-        for (int i = 0; i < (int) energyDerivs->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = energyDerivs->getParameterInfos()[i];
-            string index = cc.intToString(i+1);
-            extraArgs << ", GLOBAL " << buffer.getType() << "* RESTRICT derivBuffers" << index;
-            compute << buffer.getType() << " deriv" << index << " = derivBuffers" << index << "[index];\n";
-        }
-        if (needEnergyParamDerivs) {
-            extraArgs << ", GLOBAL mixed* RESTRICT energyParamDerivs";
-            const vector<string>& allParamDerivNames = cc.getEnergyParamDerivNames();
-            int numDerivs = allParamDerivNames.size();
-            for (int i = 0; i < force.getNumEnergyParameterDerivatives(); i++) {
-                for (int j = 0; j < dValuedParam[i]->getParameterInfos().size(); j++)
-                    extraArgs << ", GLOBAL real* RESTRICT dValuedParam_" << j << "_" << i;
-                initParamDerivs << "mixed energyParamDeriv" << i << " = 0;\n";
-                for (int index = 0; index < numDerivs; index++)
-                    if (allParamDerivNames[index] == force.getEnergyParameterDerivativeName(i))
-                        saveParamDerivs << "energyParamDerivs[GLOBAL_ID*" << numDerivs << "+" << index << "] += energyParamDeriv" << i << ";\n";
-            }
-        }
-        map<string, string> variables;
-        variables["x"] = "pos.x";
-        variables["y"] = "pos.y";
-        variables["z"] = "pos.z";
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++)
-            variables[force.getPerParticleParameterName(i)] = "params"+params->getParameterSuffix(i, "[index]");
-        for (int i = 0; i < force.getNumGlobalParameters(); i++)
-            variables[force.getGlobalParameterName(i)] = "globals["+cc.intToString(i)+"]";
-        for (int i = 0; i < numComputedValues; i++)
-            variables[computedValueNames[i]] = "values"+computedValues->getParameterSuffix(i, "[index]");
-        if (needParameterGradient) {
-            for (int i = 1; i < numComputedValues; i++) {
-                string is = cc.intToString(i);
-                compute << "real3 dV"<<is<<"dR = make_real3(0);\n";
-                for (int j = 1; j < i; j++) {
-                    if (!isZeroExpression(valueDerivExpressions[i][j])) {
-                        map<string, Lepton::ParsedExpression> derivExpressions;
-                        string js = cc.intToString(j);
-                        derivExpressions["real dV"+is+"dV"+js+" = "] = valueDerivExpressions[i][j];
-                        compute << cc.getExpressionUtilities().createExpressions(derivExpressions, variables, functionList, functionDefinitions, "temp_"+is+"_"+js);
-                        compute << "dV"<<is<<"dR += dV"<<is<<"dV"<<js<<"*dV"<<js<<"dR;\n";
-                    }
-                }
-                map<string, Lepton::ParsedExpression> gradientExpressions;
-                if (!isZeroExpression(valueGradientExpressions[i][0]))
-                    gradientExpressions["dV"+is+"dR.x += "] = valueGradientExpressions[i][0];
-                if (!isZeroExpression(valueGradientExpressions[i][1]))
-                    gradientExpressions["dV"+is+"dR.y += "] = valueGradientExpressions[i][1];
-                if (!isZeroExpression(valueGradientExpressions[i][2]))
-                    gradientExpressions["dV"+is+"dR.z += "] = valueGradientExpressions[i][2];
-                compute << cc.getExpressionUtilities().createExpressions(gradientExpressions, variables, functionList, functionDefinitions, "gradtemp_"+is);
-            }
-            for (int i = 1; i < numComputedValues; i++)
-                compute << "force -= deriv"<<energyDerivs->getParameterSuffix(i)<<"*dV"<<i<<"dR;\n";
-        }
-        if (needEnergyParamDerivs)
-            for (int i = 0; i < numComputedValues; i++)
-                for (int j = 0; j < dValuedParam.size(); j++)
-                    compute << "energyParamDeriv"<<j<<" += deriv"<<energyDerivs->getParameterSuffix(i)<<"*dValuedParam_"<<i<<"_"<<j<<"[index];\n";
-        map<string, string> replacements;
-        replacements["PARAMETER_ARGUMENTS"] = extraArgs.str()+tableArgs.str();
-        replacements["COMPUTE_FORCES"] = compute.str();
-        replacements["INIT_PARAM_DERIVS"] = initParamDerivs.str();
-        replacements["SAVE_PARAM_DERIVS"] = saveParamDerivs.str();
-        map<string, string> defines;
-        defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-        ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customGBGradientChainRule, replacements), defines);
-        gradientChainRuleKernel = program->createKernel("computeGradientChainRuleTerms");
-    }
-    {
-        // Create the code to calculate chain rule terms as part of the default nonbonded kernel.
-
-        vector<pair<ExpressionTreeNode, string> > globalVariables;
-        for (int i = 0; i < force.getNumGlobalParameters(); i++) {
-            const string& name = force.getGlobalParameterName(i);
-            string value = "globals["+cc.intToString(i)+"]";
-            globalVariables.push_back(makeVariable(name, prefix+value));
-        }
-        vector<pair<ExpressionTreeNode, string> > variables = globalVariables;
-        map<string, string> rename;
-        ExpressionTreeNode rnode(new Operation::Variable("r"));
-        variables.push_back(make_pair(rnode, "r"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Square(), rnode), "r2"));
-        variables.push_back(make_pair(ExpressionTreeNode(new Operation::Reciprocal(), rnode), "invR"));
-        for (int i = 0; i < force.getNumPerParticleParameters(); i++) {
-            const string& name = force.getPerParticleParameterName(i);
-            variables.push_back(makeVariable(name+"1", "((real) "+prefix+"params"+params->getParameterSuffix(i, "1)")));
-            variables.push_back(makeVariable(name+"2", "((real) "+prefix+"params"+params->getParameterSuffix(i, "2)")));
-            rename[name+"1"] = name+"2";
-            rename[name+"2"] = name+"1";
-        }
-        map<string, Lepton::ParsedExpression> derivExpressions;
-        stringstream chainSource;
-        Lepton::ParsedExpression dVdR = Lepton::Parser::parse(computedValueExpressions[0], functions).differentiate("r").optimize();
-        derivExpressions["real dV0dR1 = "] = dVdR;
-        derivExpressions["real dV0dR2 = "] = dVdR.renameVariables(rename);
-        chainSource << cc.getExpressionUtilities().createExpressions(derivExpressions, variables, functionList, functionDefinitions, prefix+"temp0_");
-        if (needChainForValue[0]) {
-            if (useExclusionsForValue)
-                chainSource << "if (!isExcluded) {\n";
-            chainSource << "tempForce -= dV0dR1*" << prefix << "dEdV" << energyDerivs->getParameterSuffix(0, "1") << ";\n";
-            chainSource << "tempForce -= dV0dR2*" << prefix << "dEdV" << energyDerivs->getParameterSuffix(0, "2") << ";\n";
-            if (useExclusionsForValue)
-                chainSource << "}\n";
-        }
-        for (int i = 1; i < numComputedValues; i++) {
-            if (needChainForValue[i]) {
-                chainSource << "tempForce -= dV0dR1*" << prefix << "dEdV" << energyDerivs->getParameterSuffix(i, "1") << ";\n";
-                chainSource << "tempForce -= dV0dR2*" << prefix << "dEdV" << energyDerivs->getParameterSuffix(i, "2") << ";\n";
-            }
-        }
-        map<string, string> replacements;
-        string chainStr = chainSource.str();
-        replacements["COMPUTE_FORCE"] = chainStr;
-        string source = cc.replaceStrings(CommonKernelSources::customGBChainRule, replacements);
-        vector<ComputeParameterInfo> parameters;
-        vector<ComputeParameterInfo> arguments;
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-            string paramName = prefix+"params"+cc.intToString(i+1);
-            if (chainStr.find(paramName+"1") != chainStr.npos || chainStr.find(paramName+"2") != chainStr.npos)
-                parameters.push_back(ComputeParameterInfo(buffer.getArray(), paramName, buffer.getComponentType(), buffer.getNumComponents()));
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-            string paramName = prefix+"values"+cc.intToString(i+1);
-            if (chainStr.find(paramName+"1") != chainStr.npos || chainStr.find(paramName+"2") != chainStr.npos)
-                parameters.push_back(ComputeParameterInfo(buffer.getArray(), paramName, buffer.getComponentType(), buffer.getNumComponents()));
-        }
-        for (int i = 0; i < (int) energyDerivChain->getParameterInfos().size(); i++) {
-            if (needChainForValue[i]) { 
-                ComputeParameterInfo& buffer = energyDerivChain->getParameterInfos()[i];
-                string paramName = prefix+"dEdV"+cc.intToString(i+1);
-                parameters.push_back(ComputeParameterInfo(buffer.getArray(), paramName, buffer.getComponentType(), buffer.getNumComponents()));
-            }
-        }
-        if (globals.isInitialized()) {
-            globals.upload(globalParamValues);
-            arguments.push_back(ComputeParameterInfo(globals, prefix+"globals", "float", 1));
-        }
-        nb.addInteraction(useCutoff, usePeriodic, force.getNumExclusions() > 0, cutoff, exclusionList, source, force.getForceGroup());
-        for (auto param : parameters)
-            nb.addParameter(param);
-        for (auto arg : arguments)
-            nb.addArgument(arg);
-    }
-    info = new ForceInfo(force);
-    cc.addForce(info);
-    cc.addAutoclearBuffer(longEnergyDerivs);
-}
-
-double CommonCalcCustomGBForceKernel::execute(ContextImpl& context, bool includeForces, bool includeEnergy) {
-    ContextSelector selector(cc);
-    bool deviceIsCpu = cc.getIsCPU();
-    NonbondedUtilities& nb = cc.getNonbondedUtilities();
-    int elementSize = (cc.getUseDoublePrecision() ? sizeof(double) : sizeof(float));
-    if (!hasInitializedKernels) {
-        hasInitializedKernels = true;
-        
-        // These two kernels can't be compiled in initialize(), because the nonbonded utilities object
-        // has not yet been initialized then.
-
-        {
-            int numExclusionTiles = nb.getExclusionTiles().getSize();
-            pairValueDefines["NUM_TILES_WITH_EXCLUSIONS"] = cc.intToString(numExclusionTiles);
-            int numContexts = cc.getNumContexts();
-            int startExclusionIndex = cc.getContextIndex()*numExclusionTiles/numContexts;
-            int endExclusionIndex = (cc.getContextIndex()+1)*numExclusionTiles/numContexts;
-            pairValueDefines["FIRST_EXCLUSION_TILE"] = cc.intToString(startExclusionIndex);
-            pairValueDefines["LAST_EXCLUSION_TILE"] = cc.intToString(endExclusionIndex);
-            pairValueDefines["CUTOFF"] = cc.doubleToString(cutoff);
-            ComputeProgram program = cc.compileProgram(pairValueSrc, pairValueDefines);
-            pairValueKernel = program->createKernel("computeN2Value");
-            pairValueSrc = "";
-            pairValueDefines.clear();
-        }
-        {
-            int numExclusionTiles = nb.getExclusionTiles().getSize();
-            pairEnergyDefines["NUM_TILES_WITH_EXCLUSIONS"] = cc.intToString(numExclusionTiles);
-            int numContexts = cc.getNumContexts();
-            int startExclusionIndex = cc.getContextIndex()*numExclusionTiles/numContexts;
-            int endExclusionIndex = (cc.getContextIndex()+1)*numExclusionTiles/numContexts;
-            pairEnergyDefines["FIRST_EXCLUSION_TILE"] = cc.intToString(startExclusionIndex);
-            pairEnergyDefines["LAST_EXCLUSION_TILE"] = cc.intToString(endExclusionIndex);
-            pairEnergyDefines["CUTOFF"] = cc.doubleToString(cutoff);
-            ComputeProgram program = cc.compileProgram(pairEnergySrc, pairEnergyDefines);
-            pairEnergyKernel = program->createKernel("computeN2Energy");
-            pairEnergySrc = "";
-            pairEnergyDefines.clear();
-        }
-
-        // Set arguments for kernels.
-        
-        maxTiles = (nb.getUseCutoff() ? nb.getInteractingTiles().getSize() : 0);
-        int numAtomBlocks = cc.getPaddedNumAtoms()/32;
-        pairValueKernel->addArg(cc.getPosq());
-        pairValueKernel->addArg(cc.getNonbondedUtilities().getExclusions());
-        pairValueKernel->addArg(cc.getNonbondedUtilities().getExclusionTiles());
-        pairValueKernel->addArg(valueBuffers);
-        if (nb.getUseCutoff()) {
-            pairValueKernel->addArg(nb.getInteractingTiles());
-            pairValueKernel->addArg(nb.getInteractionCount());
-            for (int i = 0; i < 5; i++)
-                pairValueKernel->addArg(); // Periodic box size arguments are set when the kernel is executed.
-            pairValueKernel->addArg(maxTiles);
-            pairValueKernel->addArg(nb.getBlockCenters());
-            pairValueKernel->addArg(nb.getBlockBoundingBoxes());
-            pairValueKernel->addArg(nb.getInteractingAtoms());
-        }
-        else
-            pairValueKernel->addArg(numAtomBlocks*(numAtomBlocks+1)/2);
-        if (globals.isInitialized())
-            pairValueKernel->addArg(globals);
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            if (pairValueUsesParam[i]) {
-                ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-                pairValueKernel->addArg(buffer.getArray());
-            }
-        }
-        for (auto& d : dValue0dParam)
-            pairValueKernel->addArg(d);
-        for (auto& function : tabulatedFunctionArrays)
-            pairValueKernel->addArg(function);
-        perParticleValueKernel->addArg(cc.getPosq());
-        perParticleValueKernel->addArg(valueBuffers);
-        if (globals.isInitialized())
-            perParticleValueKernel->addArg(globals);
-        for (auto& buffer : params->getParameterInfos())
-            perParticleValueKernel->addArg(buffer.getArray());
-        for (auto& buffer : computedValues->getParameterInfos())
-            perParticleValueKernel->addArg(buffer.getArray());
-        for (int i = 0; i < dValuedParam.size(); i++) {
-            perParticleValueKernel->addArg(dValue0dParam[i]);
-            for (int j = 0; j < dValuedParam[i]->getParameterInfos().size(); j++)
-                perParticleValueKernel->addArg(dValuedParam[i]->getParameterInfos()[j].getArray());
-        }
-        for (auto& function : tabulatedFunctionArrays)
-            perParticleValueKernel->addArg(function);
-        pairEnergyKernel->addArg(cc.getLongForceBuffer());
-        pairEnergyKernel->addArg(cc.getEnergyBuffer());
-        pairEnergyKernel->addArg(cc.getPosq());
-        pairEnergyKernel->addArg(cc.getNonbondedUtilities().getExclusions());
-        pairEnergyKernel->addArg(cc.getNonbondedUtilities().getExclusionTiles());
-        pairEnergyKernel->addArg(); // Whether to include energy.
-        if (nb.getUseCutoff()) {
-            pairEnergyKernel->addArg(nb.getInteractingTiles());
-            pairEnergyKernel->addArg(nb.getInteractionCount());
-            for (int i = 0; i < 5; i++)
-                pairEnergyKernel->addArg(); // Periodic box size arguments are set when the kernel is executed.
-            pairEnergyKernel->addArg(maxTiles);
-            pairEnergyKernel->addArg(nb.getBlockCenters());
-            pairEnergyKernel->addArg(nb.getBlockBoundingBoxes());
-            pairEnergyKernel->addArg(nb.getInteractingAtoms());
-        }
-        else
-            pairEnergyKernel->addArg(numAtomBlocks*(numAtomBlocks+1)/2);
-        if (globals.isInitialized())
-            pairEnergyKernel->addArg(globals);
-        for (int i = 0; i < (int) params->getParameterInfos().size(); i++) {
-            if (pairEnergyUsesParam[i]) {
-                ComputeParameterInfo& buffer = params->getParameterInfos()[i];
-                pairEnergyKernel->addArg(buffer.getArray());
-            }
-        }
-        for (int i = 0; i < (int) computedValues->getParameterInfos().size(); i++) {
-            if (pairEnergyUsesValue[i]) {
-                ComputeParameterInfo& buffer = computedValues->getParameterInfos()[i];
-                pairEnergyKernel->addArg(buffer.getArray());
-            }
-        }
-        pairEnergyKernel->addArg(longEnergyDerivs);
-        if (needEnergyParamDerivs)
-            pairEnergyKernel->addArg(cc.getEnergyParamDerivBuffer());
-        for (auto& function : tabulatedFunctionArrays)
-            pairEnergyKernel->addArg(function);
-        perParticleEnergyKernel->addArg(cc.getEnergyBuffer());
-        perParticleEnergyKernel->addArg(cc.getPosq());
-        perParticleEnergyKernel->addArg(cc.getLongForceBuffer());
-        if (globals.isInitialized())
-            perParticleEnergyKernel->addArg(globals);
-        for (auto& buffer : params->getParameterInfos())
-            perParticleEnergyKernel->addArg(buffer.getArray());
-        for (auto& buffer : computedValues->getParameterInfos())
-            perParticleEnergyKernel->addArg(buffer.getArray());
-        for (auto& buffer : energyDerivs->getParameterInfos())
-            perParticleEnergyKernel->addArg(buffer.getArray());
-        for (auto& buffer : energyDerivChain->getParameterInfos())
-            perParticleEnergyKernel->addArg(buffer.getArray());
-        perParticleEnergyKernel->addArg(longEnergyDerivs);
-        if (needEnergyParamDerivs)
-            perParticleEnergyKernel->addArg(cc.getEnergyParamDerivBuffer());
-        for (auto& function : tabulatedFunctionArrays)
-            perParticleEnergyKernel->addArg(function);
-        if (needParameterGradient || needEnergyParamDerivs) {
-            gradientChainRuleKernel->addArg(cc.getPosq());
-            gradientChainRuleKernel->addArg(cc.getLongForceBuffer());
-            if (globals.isInitialized())
-                gradientChainRuleKernel->addArg(globals);
-            for (auto& buffer : params->getParameterInfos())
-                gradientChainRuleKernel->addArg(buffer.getArray());
-            for (auto& buffer : computedValues->getParameterInfos())
-                gradientChainRuleKernel->addArg(buffer.getArray());
-            for (auto& buffer : energyDerivs->getParameterInfos())
-                gradientChainRuleKernel->addArg(buffer.getArray());
-            if (needEnergyParamDerivs) {
-                gradientChainRuleKernel->addArg(cc.getEnergyParamDerivBuffer());
-                for (auto d : dValuedParam)
-                    for (auto& buffer : d->getParameterInfos())
-                        gradientChainRuleKernel->addArg(buffer.getArray());
-            }
-            for (auto& function : tabulatedFunctionArrays)
-                gradientChainRuleKernel->addArg(function);
-        }
-    }
-    if (globals.isInitialized()) {
-        bool changed = false;
-        for (int i = 0; i < (int) globalParamNames.size(); i++) {
-            float value = (float) context.getParameter(globalParamNames[i]);
-            if (value != globalParamValues[i])
-                changed = true;
-            globalParamValues[i] = value;
-        }
-        if (changed)
-            globals.upload(globalParamValues);
-    }
-    pairEnergyKernel->setArg(5, (int) includeEnergy);
-    if (nb.getUseCutoff()) {
-        setPeriodicBoxArgs(cc, pairValueKernel, 6);
-        setPeriodicBoxArgs(cc, pairEnergyKernel, 8);
-        if (maxTiles < nb.getInteractingTiles().getSize()) {
-            maxTiles = nb.getInteractingTiles().getSize();
-            pairValueKernel->setArg(11, maxTiles);
-            pairEnergyKernel->setArg(13, maxTiles);
-        }
-    }
-    pairValueKernel->execute(nb.getNumForceThreadBlocks()*nb.getForceThreadBlockSize(), nb.getForceThreadBlockSize());
-    perParticleValueKernel->execute(cc.getPaddedNumAtoms());
-    pairEnergyKernel->execute(nb.getNumForceThreadBlocks()*nb.getForceThreadBlockSize(), nb.getForceThreadBlockSize());
-    perParticleEnergyKernel->execute(cc.getPaddedNumAtoms());
-    if (needParameterGradient || needEnergyParamDerivs)
-        gradientChainRuleKernel->execute(cc.getPaddedNumAtoms());
-    return 0.0;
-}
-
-void CommonCalcCustomGBForceKernel::copyParametersToContext(ContextImpl& context, const CustomGBForce& force) {
-    ContextSelector selector(cc);
-    int numParticles = force.getNumParticles();
-    if (numParticles != cc.getNumAtoms())
-        throw OpenMMException("updateParametersInContext: The number of particles has changed");
-
-    // Record the per-particle parameters.
-
-    vector<vector<float> > paramVector(cc.getPaddedNumAtoms(), vector<float>(force.getNumPerParticleParameters(), 0));
-    vector<double> parameters;
-    for (int i = 0; i < numParticles; i++) {
-        force.getParticleParameters(i, parameters);
-        for (int j = 0; j < (int) parameters.size(); j++)
-            paramVector[i][j] = (float) parameters[j];
-    }
-    params->setParameterValues(paramVector);
-
-    // See if any tabulated functions have changed.
-
-    for (int i = 0; i < force.getNumTabulatedFunctions(); i++) {
-        string name = force.getTabulatedFunctionName(i);
-        if (force.getTabulatedFunction(i).getUpdateCount() != tabulatedFunctionUpdateCount[name]) {
-            tabulatedFunctionUpdateCount[name] = force.getTabulatedFunction(i).getUpdateCount();
-            int width;
-            vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(force.getTabulatedFunction(i), width);
-            tabulatedFunctionArrays[i].upload(f);
-        }
-    }
-
-    // Mark that the current reordering may be invalid.
-
     cc.invalidateMolecules(info, true, false);
 }
